增加位置分类列表

+++ src/main/java/com/garbage/classification/dao/base/BaseDao.java
@@ -18,7 +18,7 @@ public interface BaseDao<T> {
      * @param id 对象ID
      * @return
      */
-    public int deleteByPrimaryKey(Integer id);
+    public int deleteByPrimaryKey(Long id);
 
     /**
      * 保存对象
@@ -32,7 +32,7 @@ public interface BaseDao<T> {
      * @param id
      * @return
      */
-    T selectByPrimaryKey(Integer id);
+    T selectByPrimaryKey(Long id);
 
     /**
      * 更新操作
